Fix drawTiled overflow check
The drawTiled vertex_data overflow check uses numeric_limits<uint64_t>
as the limit, and is the only draw* function to do so, with the rest
using unit16_t. The index_data vector is initialized as a
std::vector<uint16_t>.

This results in false even when the number of vertices exceeds capacity,
so when tiling produces more than 65531 vertices, indices silently
overflow, wrap around, and render garbage.

Align to the uint16_t limit used by every other overflow check in the
file.

+++ src/graphics/renderTarget.cpp
@@ -760,7 +760,7 @@ void RenderTarget::drawTiled(const sp::Rect& rect, std::string_view texture, glm
     {
         for(int y=0; y<tile_count.y; y++)
         {
-            if(vertex_data.size() >= std::numeric_limits<uint64_t>::max() - 4)
+            if(vertex_data.size() >= std::numeric_limits<uint16_t>::max() - 4)
             {
                 if (info.texture)
                     finish(info.texture);
